refactor: Extract health reimbursement selector
